Busybox 1.20.0 squashed commit for ICS

+++ android/librpc/xdr_rec.c
@@ -422,7 +422,7 @@ xdrrec_inline (XDR *xdrs, u_int len)
       break;
 
     case XDR_DECODE:
-      if ((len <= rstrm->fbtbc) &&
+      if ((len <= (u_int) rstrm->fbtbc) &&
 	  ((rstrm->in_finger + len) <= rstrm->in_boundry))
 	{
 	  buf = (int32_t *) rstrm->in_finger;
